requirements.txt
argcomplete==3.1.6
astroid==3.0.1
bandit==1.7.5
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
dill==0.3.7
dparse==0.6.3
flake8==6.1.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
isort==5.12.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==21.3
pathspec==0.11.2
pbr==5.11.1
platformdirs==3.11.0
pycodestyle==2.11.1
pyflakes==3.1.0
Pygments==2.16.1
pylint==3.0.2
pyparsing==3.1.1
PyYAML==6.0.1
requests==2.31.0
rich==13.6.0
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
safety==2.3.5
setuptools==68.2.2
smmap==5.0.1
stevedore==5.1.0
tomlkit==0.12.1
urllib3==2.0.7
wheel==0.41.2
yamllint==1.32.0
